+++ src/main/java/com/treasurehunt/treasurehunt/db/MySQL/JdbcMysqlJ8IntegrationDemo.java
@@ -1,4 +1,4 @@
-package com.treasurehunt.treasurehunt.db.mysql;
+package com.treasurehunt.treasurehunt.db.MySQL;
 
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
