+++ src/main/java/com/treasurehunt/treasurehunt/db/MySQL/MySQLException.java
@@ -1,4 +1,4 @@
-package com.treasurehunt.treasurehunt.db.mysql;
+package com.treasurehunt.treasurehunt.db.MySQL;
 
 public class MySQLException extends RuntimeException{
     public MySQLException(String errorMessage) {
