+++ src/main/java/com/treasurehunt/treasurehunt/db/MySQL/MySQLTableCreator.java
@@ -1,4 +1,4 @@
-package com.treasurehunt.db.MySQL;
+package com.treasurehunt.treasurehunt.db.MySQL;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
